Restructure routing  - Create models.py  - Add routes.py
