+++ ovation_app/__init__.py
@@ -1,24 +1,10 @@
 from flask import Flask, render_template
 
-app = Flask(__name__)
-
-@app.route('/about')
-@app.route('/landing')
-@app.route('/')
-def landing():
-    return render_template('landing.html')
-
-@app.route('/login')
-def login():
-    return render_template('auth/login.html')
-
-@app.route('/signup')
-def signUp():
-    return render_template('auth/signup.html')
-
-@app.route('/profile')
-def profile():
-    return render_template('profile.html')
-
 def create_app():
+    app = Flask(__name__, template_folder='templates')
+    from .routes import routes
+
+    app.register_blueprint(routes, url_prefix = '/')
+
     return app
+
